# swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Beyond The Label"
host: "localhost"
basePath: "/btl"
schemes:
- "https"
- "http"
paths:
...
  /product:
    post:
      tags:
      - "Product"
      consumes:
      - "application/octet-stream"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Product create request"
        required: true
        schema:
          type: object
          required:
          - "name"
          - "label"
          - "customer_id"
          - "description"
          - "video_file_path"
          - "category_id"
          properties:
            customer_id:
              type: string
              format: uuid
            category_id:
              type: string
              format: uuid
            name:
              type: string
            label:
              type: string
            description:
              type: string
            video_file_path:
              type: string
              format: binary
            intent_file_path:
              type: string
              format: binary
            metadata:
                  type: array
                  items:
                    type: object
                    properties:
                      Key:
                        type: string
                        format: string
                      value:
                        type: string
                        format: string
      responses:
        "200":
          description: "Product created successfully"
          schema:
            type: object
            properties:
              message:
                type: string
                format: string
              data:
                type: object
                properties:
                  customer_id:
                    type: string
                    format: uuid
                  product_id:
                    type: string
                    format: uuid
                  category_id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  label:
                    type: string
                  description:
                    type: string
                  video_file_path:
                    type: string
                  intent_file_path:
                    type: string
                  image_status:
                    type: boolean
                  training_status:
                    type: integer
                    format: int64
                  is_deleted:
                    type: boolean
                  product_display_id:
                    type: string
                    format: string
                  metadata:
                        type: array
                        items:
                          type: object
                          properties:
                            Key:
                              type: string
                              format: string
                            value:
                              type: string
                              format: string
        "405":
          description: "Invalid input"
    put:
      tags:
      - "Product"
      consumes:
      - "application/octet-stream"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Product create request"
        required: true
        schema:
          type: object
          required:
          - "product_id"
          properties:
            product_id:
              type: string
              format: uuid
            label:
              type: string
            description:
              type: string
            video_file_path:
              type: string
            intent_file_path:
              type: string
            metadata:
                  type: array
                  items:
                    type: object
                    properties:
                      Key:
                        type: string
                        format: string
                      value:
                        type: string
                        format: string
      responses:
        "200":
          description: "Product created successfully"
          schema:
            type: object
            properties:
              message:
                type: string
                format: string
              data:
                type: object
                properties:
                  customer_id:
                    type: string
                    format: uuid
                  product_id:
                    type: string
                    format: uuid
                  category_id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  label:
                    type: string
                  description:
                    type: string
                  video_file_path:
                    type: string
                  intent_file_path:
                    type: string
                  image_status:
                    type: boolean
                  training_status:
                    type: integer
                    format: int64
                  is_deleted:
                    type: boolean
                  product_display_id:
                    type: string
                    format: string
                  metadata:
                        type: array
                        items:
                          type: object
                          properties:
                            Key:
                              type: string
                              format: string
                            value:
                              type: string
                              format: string
        "405":
          description: "Invalid input"
    get:
      tags:
      - "Product"
      consumes:
      - "application/octet-stream"
      produces:
      - "application/json"
      responses:
        "405":
          description: "Invalid input"
        "200":
          description: "Customer Details"
          schema:
            type: array
            items:
              type: object
              properties:
                customer_id:
                  type: string
                  format: uuid
                product_id:
                  type: string
                  format: uuid
                category_id:
                  type: string
                  format: uuid
                name:
                  type: string
                label:
                  type: string
                description:
                  type: string
                video_file_path:
                  type: string
                intent_file_path:
                  type: string
                image_status:
                  type: boolean
                training_status:
                  type: integer
                  format: int64
                is_deleted:
                  type: boolean
                product_display_id:
                  type: string
                  format: string
                metadata:
                      type: array
                      items:
                        type: object
                        properties:
                          Key:
                            type: string
                            format: string
                          value:
                            type: string
                            format: string
  /product/{id}:
    delete:
      tags:
      - "Product"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Product id to delete"
        required: true
        type: string
        format: uuid
      responses:
        "405":
          description: "Invalid input"
  /product/categories:
    get:
      tags:
      - "Product Category"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "405":
          description: "Invalid Request"
        "200":
          description: "Product categories Details"
          schema:
            type: array
            items:
              type: object
              properties:
                category_id:
                  type: string
                  format: uuid
                category:
                  type: string
                  format: string
                category_code:
                  type: string
                  format: string
                description:
                  type: string
                  format: string
...
  /product/{id}/intent:
    get:
      summary: Returns all intents for a specific product
      description: This can be used to display all intents in a list which can then be cycled through to view particular intents. Filter intents by product ID using CSV.
      tags:
      - Product
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Product id for intents"
        required: true
        type: string
      responses:
        "200":
          description: "All intents for this product"
          schema:
            type: array
            items:
              $ref: '#/definitions/Intent'
        "400":
          description: ""
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                description: >
                  error:
                   * `First Error` - No Agent found for this product!
                   * `Second Error` - No intents associated with this Agent ID!
                   * `Third Error` - There was an error while fetching an intent from DialogFlow!
                   * `Fourth Error` - There was an error while building intents response!
                enum:
                - "No Agent found for this product!"
                - "No intents associated with this Agent ID!"
                - "There was an error while fetching an intent from DialogFlow!"
                - "There was an error while building intents response!"
                default: "No Agent found for this product! No intents associated with this Agent ID! There was an error while fetching an intent from DialogFlow! There was an error while building intents response!"
    post:
      summary: Creates/Updates/Deletes intents mentioned in the CSV
      description: This will be used every time the CSV is updated. It'll create/update/delete intents mentioned in the CSV depending on the Action column
      tags:
      - Product
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Product id for intents"
        required: true
        type: string
      responses:
        "200":
          description: "Product intents updated successfully."
        "400":
          description: ""
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                description: >
                  error:
                   * `First Error` - No Agent found for this product!
                   * `Second Error` - No Intents file found at the path!
                   * `Third Error` - Product intents update failed!
                enum:
                - "No Agent found for this product!"
                - "No Intents file found at the path!"
                - "Product intents update failed!"
                default: "No Agent found for this product! No Intents file found at the path! Product intents update failed!"
  /product/{id}/intent/download:
    get:
      summary: Downloads a CSV file of product intents
      description: Calls the get intent API, writes all the intents to the csv file and prompts the user to save it.
      tags:
      - Product
      consumes:
      - "application/json"
      produces:
      - "application/csv"
      parameters:
      - name: "id"
        in: "path"
        description: "Product id for intents"
        required: true
        type: string
      responses:
        "200":
          description: A CSV file
          schema:
            type: file
        "400":
          description: There was an error while downloading intents CSV!
definitions:
...
  Part:
    type: "object"
    properties:
      text:
        type: "string"
      parameterId:
        type: "string"
  TrainingPhrase:
    type: "object"
    properties:
      id:
        type: "string"
      parts:
        type: "array"
        items:
          $ref: '#/definitions/Part'
      repeatCount:
        type: "integer"
  Intent:
    type: "object"
    properties:
      displayName:
        type: "string"
      trainingPhrases:
        type: array
        items:
          $ref: '#/definitions/TrainingPhrase'
      description:
        type: "string"
      maskedIntentId:
        type: "string"
      fulfillments:
        type: "string"